event handling in page components
