--- frontend/js/scripts/login.js
@@ -1,16 +1,0 @@
-import redirect from "/js/router.js";
-
-document
-	.getElementById("loginForm")
-	.addEventListener("submit", function (event) {
-		event.preventDefault();
-
-		const username = document.getElementById("loginUser").value;
-		const password = document.getElementById("loginPass").value;
-
-		if (username && password) {
-			route("/users");
-		} else {
-			alert("Please enter both username and password.");
-		}
-	});
